+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.05
 
 * Fix error when extracting to much files
+* Improve ngpack commands
+
 ## 0.0.4
 
 * Replace dart:io to universal_io